Construct gene set edge groups for a gene set collection (GSC) of interest

The `getGeneSetEdgeGroup` function generates the gene set edge groups in `brainclass` classification.

getGeneSetEdgeGroup(geneExpr, geneSetList, cutoff = 0.99)

Arguments

geneExpr
The ROI-by-gene data frame or matrix. Potentially the output from the `filterGeneExpr` function.

geneSetList
A gene set collection stored in a list. Potentially the output from the `filterGeneSets` function

cutoff
The hard-thresholding cutoff on edge weights.

Value

A list object of vectors, each containing labels of edges that are the most highly weighted in the corresponding gene set expression network. Each item in the returned list represents a gene set edge group. These gene set edge groups are used as feature groups in `brainclass` classification.

Examples

## NOT RUN
# data(ahba)
# data(gscv7.0)
# ahba <- filterGeneExpr(ahba)
# kegg <- filterGeneSets(geneSetList = gscv7.0$kegg,
# candidateGenes = colnames(ahba), min.size = 5, max.size = Inf)
# keggEdgeGrp <- getGeneSetEdgeGroup(geneExpr = ahba, geneSetList = kegg)